create prover building and running with new composer
