+++ cpp/src/barretenberg/plonk/composer/splitting_tmp/composer_helper/ultra_plonk_composer_helper.cpp
@@ -1,12 +1,186 @@
 #include "ultra_plonk_composer_helper.hpp"
 #include "barretenberg/proof_system/circuit_constructors/ultra_circuit_constructor.hpp"
 #include "barretenberg/proof_system/composer/permutation_helper.hpp"
+#include "barretenberg/plonk/proof_system/commitment_scheme/kate_commitment_scheme.hpp"
 
 #include <cstddef>
 #include <cstdint>
 #include <string>
 
 namespace plonk {
+
+/**
+ * @brief Computes `this.witness`, which is basiclly a set of polynomials mapped-to by strings.
+ *
+ * Note: this doesn't actually compute the _entire_ witness. Things missing: randomness for blinding both the wires and
+ * sorted `s` poly, lookup rows of the wire witnesses, the values of `z_lookup`, `z`. These are all calculated
+ * elsewhere.
+ */
+template <typename CircuitConstructor>
+void UltraPlonkComposerHelper<CircuitConstructor>::compute_witness(CircuitConstructor& circuit_constructor,
+                                                                   const size_t minimum_circuit_size)
+{
+    if (computed_witness) {
+        return;
+    }
+
+    size_t tables_size = 0;
+    size_t lookups_size = 0;
+    for (const auto& table : circuit_constructor.lookup_tables) {
+        tables_size += table.size;
+        lookups_size += table.lookup_gates.size();
+    }
+
+    const size_t filled_gates = circuit_constructor.num_gates + circuit_constructor.public_inputs.size();
+    const size_t total_num_gates = std::max(filled_gates, tables_size + lookups_size);
+
+    const size_t subgroup_size = circuit_constructor.get_circuit_subgroup_size(total_num_gates + NUM_RESERVED_GATES);
+
+    // Pad the wires (pointers to `witness_indices` of the `variables` vector).
+    // Note: the remaining NUM_RESERVED_GATES indices are padded with zeros within `compute_witness_base` (called
+    // next).
+    for (size_t i = filled_gates; i < total_num_gates; ++i) {
+        circuit_constructor.w_l.emplace_back(circuit_constructor.zero_idx);
+        circuit_constructor.w_r.emplace_back(circuit_constructor.zero_idx);
+        circuit_constructor.w_o.emplace_back(circuit_constructor.zero_idx);
+        circuit_constructor.w_4.emplace_back(circuit_constructor.zero_idx);
+    }
+
+    auto wire_polynomial_evaluations =
+        compute_witness_base(circuit_constructor, minimum_circuit_size, NUM_RANDOMIZED_GATES);
+
+    for (size_t j = 0; j < program_width; ++j) {
+        std::string index = std::to_string(j + 1);
+        circuit_proving_key->polynomial_store.put("w_" + index + "_lagrange",
+                                                  std::move(wire_polynomial_evaluations[j]));
+    }
+
+    polynomial s_1(subgroup_size);
+    polynomial s_2(subgroup_size);
+    polynomial s_3(subgroup_size);
+    polynomial s_4(subgroup_size);
+    polynomial z_lookup(subgroup_size + 1); // Only instantiated in this function; nothing assigned.
+
+    // Save space for adding random scalars in the s polynomial later.
+    // The subtracted 1 allows us to insert a `1` at the end, to ensure the evaluations (and hence coefficients) aren't
+    // all 0.
+    // See ComposerBase::compute_proving_key_base for further explanation, as a similar trick is done there.
+    size_t count = subgroup_size - tables_size - lookups_size - s_randomness - 1;
+    for (size_t i = 0; i < count; ++i) {
+        s_1[i] = 0;
+        s_2[i] = 0;
+        s_3[i] = 0;
+        s_4[i] = 0;
+    }
+
+    for (auto& table : circuit_constructor.lookup_tables) {
+        const fr table_index(table.table_index);
+        auto& lookup_gates = table.lookup_gates;
+        for (size_t i = 0; i < table.size; ++i) {
+            if (table.use_twin_keys) {
+                lookup_gates.push_back({
+                    {
+                        table.column_1[i].from_montgomery_form().data[0],
+                        table.column_2[i].from_montgomery_form().data[0],
+                    },
+                    {
+                        table.column_3[i],
+                        0,
+                    },
+                });
+            } else {
+                lookup_gates.push_back({
+                    {
+                        table.column_1[i].from_montgomery_form().data[0],
+                        0,
+                    },
+                    {
+                        table.column_2[i],
+                        table.column_3[i],
+                    },
+                });
+            }
+        }
+
+#ifdef NO_TBB
+        std::sort(lookup_gates.begin(), lookup_gates.end());
+#else
+        std::sort(std::execution::par_unseq, lookup_gates.begin(), lookup_gates.end());
+#endif
+
+        for (const auto& entry : lookup_gates) {
+            const auto components = entry.to_sorted_list_components(table.use_twin_keys);
+            s_1[count] = components[0];
+            s_2[count] = components[1];
+            s_3[count] = components[2];
+            s_4[count] = table_index;
+            ++count;
+        }
+    }
+
+    // Initialise the `s_randomness` positions in the s polynomials with 0.
+    // These will be the positions where we will be adding random scalars to add zero knowledge
+    // to plookup (search for `Blinding` in plonk/proof_system/widgets/random_widgets/plookup_widget_impl.hpp
+    // ProverPlookupWidget::compute_sorted_list_polynomial())
+    for (size_t i = 0; i < s_randomness; ++i) {
+        s_1[count] = 0;
+        s_2[count] = 0;
+        s_3[count] = 0;
+        s_4[count] = 0;
+        ++count;
+    }
+
+    circuit_proving_key->polynomial_store.put("s_1_lagrange", std::move(s_1));
+    circuit_proving_key->polynomial_store.put("s_2_lagrange", std::move(s_2));
+    circuit_proving_key->polynomial_store.put("s_3_lagrange", std::move(s_3));
+    circuit_proving_key->polynomial_store.put("s_4_lagrange", std::move(s_4));
+
+    computed_witness = true;
+}
+
+template <typename CircuitConstructor>
+UltraProver UltraPlonkComposerHelper<CircuitConstructor>::create_prover(CircuitConstructor& circuit_constructor)
+{
+    finalize_circuit(circuit_constructor);
+
+    compute_proving_key(circuit_constructor);
+    compute_witness(circuit_constructor);
+
+    UltraProver output_state(circuit_proving_key, create_manifest(circuit_constructor.public_inputs.size()));
+
+    std::unique_ptr<ProverPermutationWidget<4, true>> permutation_widget =
+        std::make_unique<ProverPermutationWidget<4, true>>(circuit_proving_key.get());
+
+    std::unique_ptr<ProverPlookupWidget<>> plookup_widget =
+        std::make_unique<ProverPlookupWidget<>>(circuit_proving_key.get());
+
+    std::unique_ptr<ProverPlookupArithmeticWidget<ultra_settings>> arithmetic_widget =
+        std::make_unique<ProverPlookupArithmeticWidget<ultra_settings>>(circuit_proving_key.get());
+
+    std::unique_ptr<ProverGenPermSortWidget<ultra_settings>> sort_widget =
+        std::make_unique<ProverGenPermSortWidget<ultra_settings>>(circuit_proving_key.get());
+
+    std::unique_ptr<ProverEllipticWidget<ultra_settings>> elliptic_widget =
+        std::make_unique<ProverEllipticWidget<ultra_settings>>(circuit_proving_key.get());
+
+    std::unique_ptr<ProverPlookupAuxiliaryWidget<ultra_settings>> auxiliary_widget =
+        std::make_unique<ProverPlookupAuxiliaryWidget<ultra_settings>>(circuit_proving_key.get());
+
+    output_state.random_widgets.emplace_back(std::move(permutation_widget));
+    output_state.random_widgets.emplace_back(std::move(plookup_widget));
+
+    output_state.transition_widgets.emplace_back(std::move(arithmetic_widget));
+    output_state.transition_widgets.emplace_back(std::move(sort_widget));
+    output_state.transition_widgets.emplace_back(std::move(elliptic_widget));
+    output_state.transition_widgets.emplace_back(std::move(auxiliary_widget));
+
+    std::unique_ptr<KateCommitmentScheme<ultra_settings>> kate_commitment_scheme =
+        std::make_unique<KateCommitmentScheme<ultra_settings>>();
+
+    output_state.commitment_scheme = std::move(kate_commitment_scheme);
+
+    return output_state;
+}
 
 template <typename CircuitConstructor>
 std::shared_ptr<proving_key> UltraPlonkComposerHelper<CircuitConstructor>::compute_proving_key(
@@ -22,6 +196,10 @@ std::shared_ptr<proving_key> UltraPlonkComposerHelper<CircuitConstructor>::compu
         tables_size += table.size;
         lookups_size += table.lookup_gates.size();
     }
+
+    info("fake proving key tables_size = ", tables_size);
+    info("fake proving key lookups_size = ", lookups_size);
+    info("fake proving key num_gates = ", circuit_constructor.num_gates);
 
     const size_t minimum_circuit_size = tables_size + lookups_size;
     const size_t num_randomized_gates = NUM_RANDOMIZED_GATES;
